+++ src/main/com/portalmedia/embarc/gui/mxf/CoreMXFController.java
@@ -175,11 +175,15 @@ public class CoreMXFController extends AnchorPane {
 			selectedFilesLabel.setText(Integer.toString(num) + " files selected");
 		}
 	}
-	
+
 	public void setSection(Boolean resetValues) {
 		if (resetValues) {}
 
 		final MXFSelectedFilesSummary summary = ControllerMediatorMXF.getInstance().getSelectedFilesSummary();
+
+		if (summary.getFilesAreMissingAS07CoreDMSFramework()) {
+			toggleEditButton.setDisable(true);
+		}
 
 		for (final MXFColumn col : MXFColumn.values()) {
 			if (col.getSection() != MXFSection.CORE) continue;
@@ -228,7 +232,9 @@ public class CoreMXFController extends AnchorPane {
 			final ASCIIArea area = new ASCIIArea();
 			area.setMXFColumn(col);
 			area.setVisible(true);
-			area.setLabel(col.getDisplayName(), MXFColumnHelpText.getInstance().getHelpText(col));
+			String label = col.getDisplayName();
+			if (col.isRequired()) label += " *";
+			area.setLabel(label, MXFColumnHelpText.getInstance().getHelpText(col));
 			MetadataColumnDef columnDef = summary.getCoreData().get(col);
 			if (columnDef != null && columnDef.getCurrentValue() != null) {
 				area.setValue(columnDef.getCurrentValue());
@@ -279,7 +285,9 @@ public class CoreMXFController extends AnchorPane {
 		dropDownField.setEditable(col.getEditable());
 		dropDownField.setMXFColumn(col);
 		dropDownField.setVisible(true);
-		dropDownField.setLabel(col.getDisplayName(), MXFColumnHelpText.getInstance().getHelpText(col));
+		String label = col.getDisplayName();
+		if (col.isRequired()) label += " *";
+		dropDownField.setLabel(label, MXFColumnHelpText.getInstance().getHelpText(col));
 		MetadataColumnDef columnDef = summary.getCoreData().get(col);
 		if (columnDef != null) {
 			dropDownField.setValue(columnDef.getCurrentValue());
@@ -300,13 +308,15 @@ public class CoreMXFController extends AnchorPane {
 		DeviceSetHelper deviceSetHelper = new DeviceSetHelper();
 		ArrayList<AS07CoreDMSDeviceObjectsImpl> devices = deviceSetHelper.createDeviceListFromString(summary.getCoreData().get(col).getCurrentValue());
 		HBox hbox = new HBox();
-		Label devicesLabel = new Label("Devices");
+		String label = "Devices";
+		if (col.isRequired()) label += " *";
+		Label devicesLabel = new Label(label);
 		devicesLabel.setLayoutX(14.0);
 		devicesLabel.setLayoutY(14.0);
 		devicesLabel.setPrefHeight(26.0);
-		devicesLabel.setPrefWidth(200.0);
+		devicesLabel.setPrefWidth(225.0);
 		devicesLabel.setMinWidth(100.0);
-		devicesLabel.setMaxWidth(200.0);
+		devicesLabel.setMaxWidth(225.0);
 		HBox.setHgrow(devicesLabel, Priority.ALWAYS);
 		hbox.getChildren().add(devicesLabel);
 		editableFieldsVBox.getChildren().add(hbox);
@@ -508,7 +518,9 @@ public class CoreMXFController extends AnchorPane {
 		IdentifierSetHelper idSetHelper = new IdentifierSetHelper();
 		ArrayList<AS07DMSIdentifierSetImpl> identifiers = idSetHelper.createIdentifierListFromString(summary.getCoreData().get(col).getCurrentValue());
 		HBox hbox = new HBox();
-		Label identifierLabel = new Label("Identifiers");
+		String label = "Identifiers";
+		if (col.isRequired()) label += " *";
+		Label identifierLabel = new Label(label);
 		final Tooltip tt = new Tooltip("Identifiers" + "\n\n" + MXFColumnHelpText.getInstance().getHelpText(col));
 		tt.setStyle("-fx-text-fill: white; -fx-font-size: 12px");
 		tt.setPrefWidth(500);
@@ -518,9 +530,9 @@ public class CoreMXFController extends AnchorPane {
 		identifierLabel.setLayoutX(14.0);
 		identifierLabel.setLayoutY(14.0);
 		identifierLabel.setPrefHeight(26.0);
-		identifierLabel.setPrefWidth(200.0);
+		identifierLabel.setPrefWidth(225.0);
 		identifierLabel.setMinWidth(100.0);
-		identifierLabel.setMaxWidth(200.0);
+		identifierLabel.setMaxWidth(225.0);
 		HBox.setHgrow(identifierLabel, Priority.ALWAYS);
 		hbox.getChildren().add(identifierLabel);
 		editableFieldsVBox.getChildren().add(hbox);