+++ src/main/com/portalmedia/embarc/gui/mxf/WriteFilesDialogMXF.java
@@ -4,8 +4,10 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.LinkedList;
 import java.util.List;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -13,6 +15,7 @@ import com.portalmedia.embarc.gui.Main;
 import com.portalmedia.embarc.gui.ProgressDialog;
 import com.portalmedia.embarc.gui.helper.MXFFileList;
 import com.portalmedia.embarc.parser.FileInformation;
+import com.portalmedia.embarc.parser.mxf.MXFFileWriteResult;
 import com.portalmedia.embarc.parser.mxf.MXFMetadata;
 import com.portalmedia.embarc.parser.mxf.MXFService;
 import com.portalmedia.embarc.parser.mxf.MXFServiceImpl;
@@ -20,7 +23,7 @@ import com.portalmedia.embarc.report.HashReportValue;
 
 import javafx.concurrent.Task;
 import javafx.scene.control.Dialog;
-import javafx.scene.control.Label;
+import javafx.scene.text.Text;
 
 /**
  * Dialog pane that allows user to write edited files, displays results
@@ -35,6 +38,7 @@ public class WriteFilesDialogMXF extends Dialog {
 	boolean writeEditedOnly = false;
 	int success = 0;
 	int failures = 0;
+	Map<String, Exception> exceptions = new LinkedHashMap<String, Exception>();
 	final List<HashReportValue> report = new LinkedList<>();
 	List<FileInformation<MXFMetadata>> fileListCopy = new ArrayList<FileInformation<MXFMetadata>>();
 	private static final Logger LOGGER = Logger.getLogger(Main.class.getClass().getName());
@@ -85,11 +89,12 @@ public class WriteFilesDialogMXF extends Dialog {
 
 						try {
 							MXFService mxfService = new MXFServiceImpl(inputPath);
-							boolean isSuccess = mxfService.writeFile(outputPath, file.getFileData().getCoreColumns());
-							if (isSuccess) {
+							MXFFileWriteResult result = mxfService.writeFile(outputPath, file.getFileData().getCoreColumns());
+							if (result.isSuccess()) {
 								success++;
 							} else {
 								failures++;
+								exceptions.put(exceptions.size() + 1 + ". " + file.getName(), result.getException());
 							}
 						} catch (IOException e) {
 							LOGGER.log(Level.SEVERE, e.toString(), e);
@@ -112,7 +117,7 @@ public class WriteFilesDialogMXF extends Dialog {
 		progressDialog.activateProgressBar(task);
 		progressDialog.getDialogAlert().show();
 
-		final Label filesToProcess = new Label("Files to Write: " + Integer.toString(fileListCopy.size()));
+		final Text filesToProcess = new Text("Files to Write: " + Integer.toString(fileListCopy.size()));
 		progressDialog.setCountLabel(filesToProcess);
 
 		progressDialog.getDialogAlert().setOnCloseRequest(e -> {
@@ -125,20 +130,21 @@ public class WriteFilesDialogMXF extends Dialog {
 		new Thread(task).start();
 
 		task.setOnSucceeded(e -> {
-			final Label filesWritten = new Label("File Writing Complete!");
+			final Text filesWritten = new Text("File Writing Complete!");
 			progressDialog.setCountLabel(filesWritten);
-			final List<Label> labels = new ArrayList<>();
-			labels.add(new Label("Total Files Written: " + success));
-			labels.add(new Label("Total Failures: " + failures));
+			final List<Text> labels = new ArrayList<>();
+			labels.add(new Text("Total Files Written: " + success));
+			labels.add(new Text("Total Failures: " + failures));
 			progressDialog.showLabels(labels);
+			progressDialog.showExceptions(exceptions);
 			progressDialog.showCloseButton();
 		});
 
 		task.setOnCancelled(e -> {
-			final Label filesWritten = new Label("WRITE FILES CANCELLED");
+			final Text filesWritten = new Text("WRITE FILES CANCELLED");
 			progressDialog.setCountLabel(filesWritten);
-			final List<Label> labels = new ArrayList<>();
-			labels.add(new Label("No edited files were found."));
+			final List<Text> labels = new ArrayList<>();
+			labels.add(new Text("No edited files were found."));
 			progressDialog.showLabels(labels);
 			progressDialog.cancelProgressBar();
 			progressDialog.showCloseButton();