+++ src/main/com/portalmedia/embarc/parser/ColumnDef.java
@@ -28,4 +28,5 @@ public interface ColumnDef {
 
 	int getSortOrder();
 
+	boolean isRequired();
 }